\name{maqcExpression4plex-package}
\alias{maqcExpression4plex-package}
\alias{maqcExpression4plex}
\docType{package}
\title{
  Sample Expression Data - MAQC / HG18 - NimbleGen
}
\description{

  Sample dataset provided by NimbleGen Systems from the human (HG18)
  4-plex array, which has 24K genes with 3 60-mer probes/gene.
  
}

\details{
\tabular{ll}{
Package: \tab maqcExpression4plex\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab April/2007\cr
License: \tab GPL\cr
}

phenoData information in extdata/phenoData.txt

The maintainer does not warrant the accuracy of the data.

}
\author{
  Data generated by NimbleGen Systems.

  Package maintained by Benilton Carvalho <Benilton.Carvalho@cancer.org.uk>.
}

\keyword{ package }
\examples{
## None yet.
}
